pyproject.toml
[tool.poetry]
name = "youtube"
version = "0.1.0"
description = ""
authors = ["russarr <russarr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-auth = "^2.33.0"
google-auth-oauthlib = "^1.2.1"
google-api-python-client = "^2.140.0"
google-api-python-client-stubs = "^1.26.0"
pydantic = "^2.8.2"
pymongo = "^4.8.0"
pyyaml = "^6.0.2"
lxml = "^5.3.0"
lxml-stubs = "^0.5.1"
aiohttp = "^3.10.5"
isodate = "^0.6.1"
jinja2 = "^3.1.4"
httpx = "^0.27.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.basedpyright]
reportMissingTypeStubs = false
reportUnknownMemberType = false
